PYIC-7556: refactor remainding tests
